PICDataStructures

PICDataStructures.AbstractAxisGrid

PICDataStructures.AxisGrid

PICDataStructures.SparseAxisGrid

PICDataStructures.autonames

PICDataStructures.downsample

PICDataStructures.find_field

PICDataStructures.AbstractAxisGrid — Type

AbstractAxisGrid{N,T,Names}

An AbstractAxisGrid describes a structured grid in N dimensions. Each axis has a name (form Names) and the type of the elements is T.

PICDataStructures.AxisGrid — Type

AxisGrid{N,Nm1,T,R,Names}

This is a N dimensional representation of a grid discretization which has a structure where the discretization along each axis can be defined through an AbstractVector. This implies that the discretization step along each of the axes is not constant.

PICDataStructures.SparseAxisGrid — Type

SparseAxisGrid{N,Nm1,T,R,Names}

This is a N dimensional representation of a grid discretization which has a regular structure where the discretization along each axis can be defined through an AbstractRange. This implies that the grid can be defined only through the start, stop and step for each axis.

PICDataStructures.autonames — Method

autoname(N)

Given a N generate a tuple of length N with :x,:y,:z or :x1,...:xn

PICDataStructures.downsample — Method

downsample(f, target_size...)

Downsample the given input f to target_size. If no size is provided, one is computed based on the type.

PICDataStructures.find_field — Method

A = find_filed(Fs) returns the first AbstractPICDataStructure among the arguments.