PICDataStructures

PICDataStructures.AbstractAxisGrid

PICDataStructures.AxisGrid

PICDataStructures.SparseAxisGrid

PICDataStructures.autonames

PICDataStructures.axisnames

PICDataStructures.axisnames

PICDataStructures.downsample

PICDataStructures.find_field

PICDataStructures.plotdata

PICDataStructures.AbstractAxisGrid — Type

AbstractAxisGrid{N,T,Names}

An AbstractAxisGrid describes a structured grid in N dimensions. Each axis has a name (form Names) and the type of the elements is T.

PICDataStructures.AxisGrid — Type

AxisGrid{N,Nm1,T,R,Names}

This is a N dimensional representation of a grid discretization which has a structure where the discretization along each axis can be defined through an AbstractVector. This implies that the discretization step along each of the axes is not constant.

PICDataStructures.SparseAxisGrid — Type

SparseAxisGrid{N,Nm1,T,R,Names}

This is a N dimensional representation of a grid discretization which has a regular structure where the discretization along each axis can be defined through an AbstractRange. This implies that the grid can be defined only through the start, stop and step for each axis.

PICDataStructures.autonames — Method

autoname(N)

Given a N generate a tuple of length N with :x,:y,:z or :x1,...:xn

PICDataStructures.axisnames — Method

axisnames(grid::AbstractGrid; include_units=true)

Get the names of the axis of the input grid. In the case of Unitful quantities, they can be excluded by setting include_units to false.

PICDataStructures.axisnames — Method

axisnames(f::AbstractPICDataStructure; include_units=true)

Get the names of the axis of the grid corresponding to the input data structure f. In the case of Unitful quantities, they can be excluded by setting include_units to false.

PICDataStructures.downsample — Method

downsample(f, target_size...)

Downsample the given input f to target_size. If no size is provided, one is computed based on the type.

PICDataStructures.find_field — Method

A = find_filed(Fs) returns the first AbstractPICDataStructure among the arguments.

PICDataStructures.plotdata — Method

plotdata(f::AbstractPICDataStructure; kwargs...)

Universal plotting function for AbstractPICDataStructures. It uses fieldplot or scattervariable depending of the input type of f and adds a colorbar.

Keyword arguments

fig: The Figure in which to add the plot. By default it creates a new figure with Figure()

figurepos: The FigurePosition to use. The default is fig[1,1]

xlabel: The labes of the x axis of the figure. By default is determined with axisnames

ylabel: The labes of the y axis of the figure. By default is determined with axisnames

zlabel: The labes of the z axis of the figure. By default is determined with axisnames

title: The figure title. Empty ("") by default

downsample_size: Thw approximate downsample size for the input data.

By default (:default), it uses the heuristic values in downsample. If you want to avoid downsampling the data, use nothing. Otherwise, the argument is the approximate maiximun size along any direction.

cbar_orientation: The orientation of the attached Colorbar. By default is :vertical.

cbar_label: The Colorbar label. Empty ("") by default.

saveplot: Whether to save the plot or not. Default is false.

filename: The filename to use if saving the plot.

Any additional Keyword arguments are passed to the internal plotting function (fieldplot or scattervariable).